--- Minecraft/Assets/3.Script/Kangmin/Inventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39790a45e348f3045815d274571e95cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
