+++ Minecraft/Assets/3.Script/Kangmin/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63807585b33337f4cb53f10e5709d18b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
