--- Minecraft/Assets/3.Script/Kangmin/InventoryData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c9b5b6e65f81304daea4a4fe217ce41
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
