+++ Minecraft/Assets/3.Script/Kangmin/Option/Option.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4e27838a2077ec2449df69b685fcda4a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
