+++ Minecraft/Assets/Resources/Kangmin/Font.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 13072e21fec118941a54cd98cf45d025
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
